# scripts/maintenance/openapi_baseline.json
{
  "components": {
    "schemas": {
...
      "BookDetailOut": {
        "properties": {
          "asin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asin"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "character_count": {
            "title": "Character Count",
            "type": "integer"
          },
          "discussion_count": {
            "title": "Discussion Count",
            "type": "integer"
          },
          "indexed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "is_indexed": {
            "title": "Is Indexed",
            "type": "boolean"
          },
          "isbn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ocr_done_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Done At"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id"
          },
          "series_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Index"
          },
          "series_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "summary_generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Generated At"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "name",
          "authors",
          "series_id",
          "series_title",
          "is_indexed",
          "page_count",
          "indexed_at",
          "thumbnail_url",
          "ocr_done_at",
          "volume",
          "publisher",
          "asin",
          "series_index",
          "isbn",
          "summary",
          "summary_generated_at",
          "character_count",
          "discussion_count"
        ],
        "title": "BookDetailOut",
        "type": "object"
      },
...
      "CharacterDetail": {
        "properties": {
          "first_page": {
            "title": "First Page",
            "type": "integer"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "top_scenes": {
            "items": {
              "$ref": "#/components/schemas/CharacterScene"
            },
            "title": "Top Scenes",
            "type": "array"
          }
        },
        "required": [
          "name",
          "first_page",
          "page_count",
          "summary",
          "generated_at",
          "top_scenes"
        ],
        "title": "CharacterDetail",
        "type": "object"
      },
      "CharacterScene": {
        "properties": {
          "char_count": {
            "title": "Char Count",
            "type": "integer"
          },
          "page_no": {
            "title": "Page No",
            "type": "integer"
          }
        },
        "required": [
          "page_no",
          "char_count"
        ],
        "title": "CharacterScene",
        "type": "object"
      },
      "CharacterSummary": {
        "properties": {
          "first_page": {
            "title": "First Page",
            "type": "integer"
          },
          "has_summary": {
            "title": "Has Summary",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "first_page",
          "page_count",
          "has_summary"
        ],
        "title": "CharacterSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NovelMetaPatchRequest": {
        "description": "novel 1 冊のメタを部分更新するリクエスト（4.3）。省略されたフィールドは変更しない。",
        "properties": {
          "asin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asin"
          },
          "authors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authors"
          },
          "isbn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "volume_clear": {
            "default": false,
            "title": "Volume Clear",
            "type": "boolean"
          }
        },
        "title": "NovelMetaPatchRequest",
        "type": "object"
      },
      "NovelMetaUpdateResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "NovelMetaUpdateResponse",
        "type": "object"
      },
...
      "SimilarBookOut": {
        "description": "find_similar_books() の返却値に合わせる。",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "name",
          "score"
        ],
        "title": "SimilarBookOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "WebP / ZIP 画像を PDF 化してブラウザ閲覧するマルチソース（doujin / comic / novel）ライブラリのバックエンド API。Kindle キャプチャ連携、yomitoku による OCR 検索可能 PDF 生成、novel 向けの bge-m3 + Qwen によるフルテキスト検索・RAG 質問応答を含む。",
    "title": "Pic2PDF Viewer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/meta/novel/{book_key}": {
      "patch": {
        "description": "novel ソースの 1 冊メタを部分更新する（4.3）。\n\nbook_key は \"{stem}.pdf\" 形式。省略フィールドは変更しない。",
        "operationId": "patch_novel_meta_api_meta_novel__book_key__patch",
        "parameters": [
          {
            "in": "path",
            "name": "book_key",
            "required": true,
            "schema": {
              "title": "Book Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovelMetaPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelMetaUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Novel Meta",
        "tags": [
          "meta"
        ]
      }
    },
...
    "/api/novel_db/books/{book_name}": {
      "get": {
        "description": "単一書籍の詳細情報（summary / character_count / discussion_count 含む）を返す。",
        "operationId": "get_book_detail_route_api_novel_db_books__book_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "book_name",
            "required": true,
            "schema": {
              "title": "Book Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Book Detail Route",
        "tags": [
          "novel_db"
        ]
      }
    },
    "/api/novel_db/books/{book_name}/characters": {
      "get": {
        "description": "書籍に登録済みのキャラ一覧を返す（B-15）。\n\n`book_characters` に未登録（CLI 未実行）の書籍は空配列。フロントは空配列なら\n「キャラ辞典 未生成」表示にフォールバックする。",
        "operationId": "get_book_characters_api_novel_db_books__book_name__characters_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_name",
            "required": true,
            "schema": {
              "title": "Book Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CharacterSummary"
                  },
                  "title": "Response Get Book Characters Api Novel Db Books  Book Name  Characters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Book Characters",
        "tags": [
          "novel_db"
        ]
      }
    },
    "/api/novel_db/books/{book_name}/characters/{char_name}": {
      "get": {
        "description": "書籍 × キャラの詳細（サマリ + 主要シーン top 5）を返す（B-15）。",
        "operationId": "get_book_character_detail_api_novel_db_books__book_name__characters__char_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "char_name",
            "required": true,
            "schema": {
              "title": "Char Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "book_name",
            "required": true,
            "schema": {
              "title": "Book Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Book Character Detail",
        "tags": [
          "novel_db"
        ]
      }
    },
    "/api/novel_db/books/{book_name}/similar": {
      "get": {
        "description": "指定書籍に意味的に近い書籍を返す（B-19）。サマリ embedding の KNN。",
        "operationId": "get_similar_books_api_novel_db_books__book_name__similar_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_name",
            "required": true,
            "schema": {
              "title": "Book Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "top",
            "required": false,
            "schema": {
              "default": 5,
              "title": "Top",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SimilarBookOut"
                  },
                  "title": "Response Get Similar Books Api Novel Db Books  Book Name  Similar Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Similar Books",
        "tags": [
          "novel_db"
        ]
      }
    },
...
  }
}